move function to helpers; clean-up function
